pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oubliette-shield"
version = "0.4.0"
description = "AI LLM Firewall -- Runtime defense for LLM applications against prompt injection, jailbreak, and adversarial attacks"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "Oubliette Security", email = "info@oubliettesecurity.com" },
]
keywords = [
    "llm", "security", "firewall", "prompt-injection", "jailbreak",
    "ai-safety", "red-team", "honeypot", "adversarial", "defense",
    "owasp", "nist", "mitre-atlas", "cwe", "compliance", "llm-firewall",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "requests>=2.28.0",
]

[project.optional-dependencies]
flask = [
    "flask>=2.3.0",
]
ollama = [
    "ollama>=0.1.0",
]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.40.0",
]
azure = [
    "openai>=1.0.0",
]
bedrock = [
    "boto3>=1.34.0",
]
vertex = [
    "google-cloud-aiplatform>=1.40.0",
]
gemini = [
    "google-generativeai>=0.8.0",
]
llamacpp = [
    "llama-cpp-python>=0.2.0",
]
transformers = [
    "transformers>=4.35.0",
    "torch>=2.0.0",
]
langchain = [
    "langchain-core>=0.1.0",
]
llamaindex = [
    "llama-index-core>=0.10.0",
]
fastapi = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]
langgraph = [
    "langgraph>=0.2.0",
]
litellm = [
    "litellm>=1.40.0",
]
crewai = [
    "crewai>=0.50.0",
]
haystack = [
    "haystack-ai>=2.0.0",
]
semantic-kernel = [
    "semantic-kernel>=1.0.0",
]
dspy = [
    "dspy-ai>=2.5.0",
]
webhooks = [
    "requests>=2.28.0",
]
allama = [
    "requests>=2.28.0",
]
all = [
    "flask>=2.3.0",
    "ollama>=0.1.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    "google-generativeai>=0.8.0",
    "llama-cpp-python>=0.2.0",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "langchain-core>=0.1.0",
    "llama-index-core>=0.10.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]
pyrit = [
    "pyrit-core>=0.11.0",
]
deepteam = [
    "deepteam>=1.0.0",
]
redteam = [
    "pyrit-core>=0.11.0",
    "deepteam>=1.0.0",
]
dotenv = [
    "python-dotenv>=1.0.0",
]
drift = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
]
loadtest = [
    "locust>=2.20.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "locust>=2.20.0",
]

[project.urls]
Homepage = "https://github.com/oubliettesecurity/oubliette-shield"
Documentation = "https://github.com/oubliettesecurity/oubliette-shield#readme"
Repository = "https://github.com/oubliettesecurity/oubliette-shield"
Issues = "https://github.com/oubliettesecurity/oubliette-shield/issues"

[tool.setuptools.packages.find]
include = ["oubliette_shield*"]
exclude = ["tests*", "test_*", "anomaly-detection*", "AI-CTF*", "threat_intel*"]

[tool.setuptools.package-data]
oubliette_shield = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]
